Reorganize utils into modular subfolders
