+++ src/components/roster/AddPlayerDrawer.jsx
@@ -1,8 +1,8 @@
 // src/components/roster/AddPlayerDrawer.jsx
 import React, { useState, useMemo } from 'react';
 import PlayerRowMini from '@/components/roster/PlayerRowMini';
-import expandPositionGroup from '@/utils/expandPositionGroup.js';
-import { getDefaultAddPlayerFilters } from '@/utils/addPlayerUtils.js';
+import { expandPositionGroup } from '@/utils/roles';
+import { getDefaultAddPlayerFilters } from '@/utils/filtering';
 import DrawerHeader from './addPlayer/DrawerHeader';
 import PlayerSearchBar from './addPlayer/PlayerSearchBar';
 import FilterTabs from './addPlayer/FilterTabs';
