+++ src/utils/formatting/index.js
@@ -1,0 +1,2 @@
+export * from './formatHeight.js';
+export * from './formatSalary.js';
