--- src/utils/parseHeight.js
@@ -1,4 +1,0 @@
-export default function parseHeight(ht = '0-0') {
-  const parts = String(ht).split('-');
-  return parseInt(parts[0]) * 12 + parseInt(parts[1]);
-}
