--- src/utils/parseWeight.js
@@ -1,3 +1,0 @@
-export default function parseWeight(wt = 0) {
-  return parseInt(wt) || 0;
-}
