+++ src/utils/roles/positionMap.js
@@ -1,0 +1,13 @@
+export const POSITION_MAP = {
+  Guard: 'G',
+  'Point Guard': 'PG',
+  'Shooting Guard': 'SG',
+  Forward: 'F',
+  'Small Forward': 'SF',
+  'Power Forward': 'PF',
+  Center: 'C',
+  'Forward-Center': 'F/C',
+  'Guard-Forward': 'G/F',
+  'Forward-Guard': 'F',
+  'Center-Forward': 'C',
+};
